+++ src/main/java/pl/allegro/finance/tradukisto/internal/Container.java
@@ -6,10 +6,18 @@ import pl.allegro.finance.tradukisto.internal.converters.IntegerToWordsConverter
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValues;
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValuesForSmallNumbers;
+import pl.allegro.finance.tradukisto.internal.languages.english.AmericanEnglishValues;
 import pl.allegro.finance.tradukisto.internal.languages.english.EnglishValues;
+import pl.allegro.finance.tradukisto.internal.languages.french.FrenchIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.french.FrenchValues;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanThousandToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanValues;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianThousandToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianValues;
+import pl.allegro.finance.tradukisto.internal.languages.kazakh.KazakhValues;
+import pl.allegro.finance.tradukisto.internal.languages.latvian.LatvianValues;
 import pl.allegro.finance.tradukisto.internal.languages.persian.PersianValues;
 import pl.allegro.finance.tradukisto.internal.languages.polish.PolishValues;
 import pl.allegro.finance.tradukisto.internal.languages.portuguese.BrazilianPortugueseValues;
@@ -17,6 +25,14 @@ import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseInt
 import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseIntegerToWordsConverterAdapter;
 import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseThousandToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.russian.RussianValues;
+import pl.allegro.finance.tradukisto.internal.languages.serbian.SerbianValues;
+import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValues;
+import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValuesForSmallNumbers;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishBigDecimalToBankingMoneyConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishSmallNumbersToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishValues;
+import pl.allegro.finance.tradukisto.internal.languages.ukrainian.UkrainianValues;
 
 public final class Container {
 
@@ -26,6 +42,14 @@ public final class Container {
 
     public static Container russianContainer() {
         return new Container(new RussianValues());
+    }
+
+    public static Container serbianContainer() {
+        return new Container(new SerbianValues());
+    }
+
+    public static Container ukrainianContainer() {
+        return new Container(new UkrainianValues());
     }
 
     public static Container czechContainer() {
@@ -44,8 +68,42 @@ public final class Container {
         return new Container(integerConverter, bigDecimalBankingMoneyValueConverter);
     }
 
+    public static Container slovakContainer() {
+        SlovakValues slovakValues = new SlovakValues();
+        Container containerForBigNumbers = new Container(slovakValues);
+        Container containerForSmallNumbers = new Container(new SlovakValuesForSmallNumbers());
+
+        IntegerToStringConverter integerConverter = new CzechIntegerToWordsConverter(containerForBigNumbers.getNumbersConverter(),
+                containerForSmallNumbers.getNumbersConverter(), slovakValues.exceptions());
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(integerConverter,
+                slovakValues.currency());
+
+        return new Container(integerConverter, bigDecimalBankingMoneyValueConverter);
+    }
+
     public static Container englishContainer() {
         return new Container(new EnglishValues());
+    }
+
+    public static Container americanEnglishContainer() {
+        return new Container(new AmericanEnglishValues());
+    }
+
+    public static Container frenchContainer() {
+
+        FrenchValues values = new FrenchValues();
+
+        HundredsToWordsConverter hundredsToWordsConverter =
+                new HundredsToWordsConverter(values.baseNumbers(), values.twoDigitsNumberSeparator());
+        IntegerToWordsConverter frenchIntegerToWordsConverter =
+                new IntegerToWordsConverter(hundredsToWordsConverter, values.pluralForms());
+        IntegerToStringConverter converter =
+                new FrenchIntegerToWordsConverter(frenchIntegerToWordsConverter, values.exceptions(), values.pluralForms());
+
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter =
+                new BigDecimalToBankingMoneyConverter(converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
     }
 
     public static Container germanContainer() {
@@ -63,6 +121,26 @@ public final class Container {
                 converter, values.currency());
 
         return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
+    public static Container italianContainer() {
+        ItalianValues values = new ItalianValues();
+
+        ItalianThousandToWordsConverter italianThousandToWordsConverter = new ItalianThousandToWordsConverter(
+                values.baseNumbers());
+
+        IntegerToStringConverter converter = new ItalianIntegerToWordsConverter(
+                new IntegerToWordsConverter(italianThousandToWordsConverter, values.pluralForms()), values.exceptions(),
+                italianThousandToWordsConverter);
+
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(
+                converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
+    public static Container latvianContainer() {
+        return new Container(new LatvianValues());
     }
 
     public static Container brazilianPortugueseContainer() {
@@ -96,6 +174,24 @@ public final class Container {
                         persianValues.currency()));
     }
 
+
+    public static Container turkishContainer() {
+        TurkishValues values = new TurkishValues();
+
+        TurkishSmallNumbersToWordsConverter smallNumbersConverter = new TurkishSmallNumbersToWordsConverter(values);
+        IntegerToWordsConverter bigNumbersConverter = new IntegerToWordsConverter(smallNumbersConverter, values.pluralForms());
+        IntegerToStringConverter converter =
+                new TurkishIntegerToWordsConverter(bigNumbersConverter, smallNumbersConverter);
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter =
+                new TurkishBigDecimalToBankingMoneyConverter(converter, values);
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
+    public static Container kazakhContainer() {
+        KazakhValues kazakhValues = new KazakhValues();
+        return new Container(kazakhValues);
+    }
 
     private final IntegerToStringConverter integerConverter;
     private final BigDecimalToStringConverter bigDecimalConverter;