+++ src/main/java/pl/allegro/finance/tradukisto/internal/languages/MissingFormException.java
@@ -1,6 +1,7 @@
 package pl.allegro.finance.tradukisto.internal.languages;
 
 public class MissingFormException extends RuntimeException {
+
     MissingFormException(String message) {
         super(message);
     }