+++ src/test/groovy/pl/allegro/finance/tradukisto/ValueConvertersTest.groovy
@@ -3,23 +3,23 @@ package pl.allegro.finance.tradukisto
 import com.google.common.base.VerifyException
 import spock.lang.Specification
 
-import static pl.allegro.finance.tradukisto.ValueConverters.CZECH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.ENGLISH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.GERMAN_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.BRAZILIAN_PORTUGUESE_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.POLISH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.RUSSIAN_INTEGER
+import static pl.allegro.finance.tradukisto.ValueConverters.*
 
 class ValueConvertersTest extends Specification {
 
     def "should convert numbers in Brazilian Portuguese"() {
         expect:
         BRAZILIAN_PORTUGUESE_INTEGER.asWords(1_234) == "mil duzentos e trinta e quatro"
-    }	
+    }
 
     def "should convert numbers in German"() {
         expect:
         GERMAN_INTEGER.asWords(1_234) == "eintausendzweihundertvierunddreißig"
+    }
+
+    def "should convert numbers in Italian"() {
+        expect:
+        ITALIAN_INTEGER.asWords(1_234) == "milleduecentotrentaquattro"
     }
 
     def "should convert numbers in Russian"() {
@@ -40,6 +40,26 @@ class ValueConvertersTest extends Specification {
     def "should convert numbers in English"() {
         expect:
         ENGLISH_INTEGER.asWords(1_234) == "one thousand two hundred thirty-four"
+    }
+
+    def "should convert numbers in Latvian"() {
+        expect:
+        LATVIAN_INTEGER.asWords(1_234) == "viens tūkstotis divi simti trīsdesmit četri"
+    }
+
+    def "should convert numbers in Kazakh"() {
+        expect:
+        KAZAKH_INTEGER.asWords(1_234) == "бір мың екі жүз отыз төрт"
+    }
+
+    def "should convert numbers in Turkish"() {
+        expect:
+        TURKISH_INTEGER.asWords(1_234) == "Bin İki Yüz Otuz Dört"
+    }
+
+    def "should convert numbers in French"() {
+        expect:
+        FRENCH_INTEGER.asWords(1_234) == "mille deux cent trente-quatre"
     }
 
     def "should throw exception when null given"() {