+++ src/test/groovy/pl/allegro/finance/tradukisto/internal/languages/turkish/TurkishPluralFormsTest.groovy
@@ -1,0 +1,16 @@
+package pl.allegro.finance.tradukisto.internal.languages.turkish
+
+import spock.lang.Specification
+
+class TurkishPluralFormsTest extends Specification {
+
+    def pluralForms = new TurkishPluralForms("Bin")
+
+    def "should provide value for singular form"() {
+        expect:
+        pluralForms.formFor(value) == "Bin"
+
+        where:
+        value << (0..100)
+    }
+}