+++ CHANGELOG.md
@@ -27,7 +27,7 @@ Added support for:
 
 ## 0.4.1 (2015-07-22)
 
-Fixed currency symbol for Russian
+Fixed currency symbol for Russian language
 
 ## 0.5.0 (2015-09-29)
 
@@ -37,5 +37,82 @@ Added support for:
 
 ## 0.5.1 (2015-09-30)
 
-Fixed incorrect translation of "2" for Czech
+Fixed incorrect translation of "2" for Czech language
 
+## 1.0.0 (2017-09-06)
+
+Migrate to Java 8. Added support for:
+
+  - Brazilian Portuguese
+
+## 1.0.1 (2018-02-02)
+
+Fixed incorrect translations for Russian language
+
+## 1.1.0 (2018-03-26)
+
+Added support for:
+
+  - Latvian
+
+## 1.2.0 (2018-03-31)
+
+Added support for:
+
+  - Slovak
+
+## 1.3.0 (2018-05-14)
+
+Added support for:
+
+  - Kazakh
+
+## 1.4.0 (2018-06-04)
+
+Added support for:
+
+  - Ukrainian
+
+## 1.4.1 (2018-06-06)
+
+Added missing Ukrainian banking money value.
+
+## 1.5.0 (2018-09-18)
+
+Added support for:
+
+  - French
+
+## 1.6.0 (2019-01-02)
+
+Added support for:
+
+  - Serbian (latin)
+  
+Fixed:
+
+  - Incorrect translations for French language
+
+## 1.7.0 (2019-02-04)
+
+Added support for:
+
+  - Italian
+
+## 1.7.1 (2019-02-07)
+
+Fixed:
+
+  - Missing French banking money value
+
+## 1.8.0 (2019-03-05)
+
+Added support for:
+
+  - Turkish
+
+## 1.9.0 (2021-09-29)
+
+Migrated builds to GitHub Actions. Added support for:
+
+  - American English ($ instead of £ currency symbol)